requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
crispy-bootstrap5==0.6
dj-database-url==0.5.0
Django==4.0
django-crispy-forms==1.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
docopt==0.6.2
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
pipreqs==0.4.11
psycopg2==2.9.3
python-dotenv==0.19.2
python-http-client==3.3.7
pytz==2022.5
random-password-generator==2.2.0
requests==2.27.1
sendgrid==6.9.3
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
XlsxWriter==3.0.3
yarg==0.1.9
django-filter~=22.1
environ~=1.0
django-cors-headers==3.10.0
PyJWT~=2.6.0
django-filter~=22.1
google-cloud-storage
google~=3.0.0